Update QueryStats to be able to handle large datasets (#508)

+++ ClickHouse.Client/ADO/QueryStats.cs
@@ -1,5 +1,5 @@
 ﻿namespace ClickHouse.Client.ADO;
 
 #pragma warning disable SA1313 // Parameter names should begin with lower-case letter
-public record class QueryStats(int ReadRows, int ReadBytes, int WrittenRows, int WrittenBytes, int TotalRowsToRead);
+public record class QueryStats(long ReadRows, long ReadBytes, long WrittenRows, long WrittenBytes, long TotalRowsToRead);
 #pragma warning restore SA1313 // Parameter names should begin with lower-case letter
